refactor: move Scope to where it's used
